pyproject.toml
[project]
name = "llm-ocr"
version = "0.1.0"
description = ""
authors = [
    { name = "Ion Jaureguialzo Sarasola" }
]
readme = "README.md"
requires-python = "^3.14"
dependencies = [
    "pymupdf>=1.24.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
